rewrite pronouns tester and suggest-examples commands
