--- bot/libs/cog_utils/commons/__init__.py
@@ -1,1 +1,0 @@
-from .utils import register_user as register_user
