--- bot/libs/cog_utils/pronouns/__init__.py
@@ -1,10 +1,0 @@
-from .parse_utils import (
-    convert_to_proper_name as convert_to_proper_name,
-    convert_to_proper_sentence as convert_to_proper_sentence,
-    parse_pronouns_sentence as parse_pronouns_sentence,
-    validate_for_templating as validate_for_templating,
-)
-from .utils import (
-    build_approve_embed as build_approve_embed,
-    parse_pronouns as parse_pronouns,
-)
