+++ bot/libs/ui/pronouns/__init__.py
@@ -1,12 +1,7 @@
-from .modals import PronounsTesterModal as PronounsTesterModal
 from .profile_pages import PronounsProfilePages as PronounsProfilePages
 from .structs import (
     PronounsProfileCircleEntry as PronounsProfileCircleEntry,
     PronounsProfileEntry as PronounsProfileEntry,
     PronounsValuesEntry as PronounsValuesEntry,
     PronounsWordsEntry as PronounsWordsEntry,
-)
-from .views import (
-    ApprovePronounsExampleView as ApprovePronounsExampleView,
-    SuggestionView as SuggestionView,
 )
